Transaction endpoint

+++ src/main/kotlin/ar/edu/unq/criptop2p/controller/dto/TotalTransactionReportDTO.kt
@@ -2,8 +2,8 @@ package ar.edu.unq.criptop2p.controller.dto
 
 import java.util.*
 
-class TotalTransactionReportDTO ( val totalUSD:Long,
-                                  val totalARG:Long,
+class TotalTransactionReportDTO ( val totalUSD:Double,
+                                  val totalARG:Double,
                                   val currencyDetail:List<CurrencyTransactionReportDTO>,
                                   val timestamp:Date = Date()
 )
